zhipu_chat_demo/requirements.txt
openai
zai-sdk
